Remove operator declaration
I did this because it was causing the compiler to see >>> as an
ambiguous operator when both Functional & ButtonStyleKit were being
imported.

This change will now require that consumers first import Functional and
then import ButtonStyleKit as follows:

import Functional
import ButtonStyleKit

This is because Functional declares the >>> operator and ButtonStyleKit
overrides the >>> operator for use with the composable button styling
provided by ButtonStyleKit.